Use official grolifant release
Description
===========

This patch removes the custom code pushed to the plugin because
grolifant had no mavencentral release. This has changed for quite some
time now. I removed all local grolifant sources and implemented
the downloader etc with the latest version.
I want to treat this as a breaking change because I changed quite a lot.
I also pulled in the latest version of the plugins plugin which will
result in a groupname change on the gradle plugins portal.

Changes
=======

* ![REMOVE] local grolifant classes
* ![UPDATE] rust downloader etc with latest grolifant release
* ![UPDATE] `net.wooga.plugins` to version 3.x