Add hooks to set cargo home and target
Description
===========

In order to use this plugin in combination with the new `net.wooga.rustup` plugin
we need some minor additions. I tried to keep this at a minimum because I think
a bigger refactor for future proofing this plugin is in need and I only wanted to
add the support to use cargo from an external location. That was actually already possible
but I found out that without the setting of `CARGO_HOME` some weird path issues can happen.
That's why we pass the `CARGO_HOME` variable to the cargo execution.

I also added a way to set the build target tripplet in the base extension which
gets pulled from environment or gradle properties

Changes
=======

* ![IMPROVE] `CARGO_HOME` setup in the compile task
* ![IMPROVE] pass `target` property from extension to task
* ![IMPROVE] expose `target` property from extension to env/properties