Fix cargo compile task output dir based on selected target
Description
===========

The `outputDir` property was declared with a semi static value for
the output location of the compiled files. Cargo will put the output
for specificly selected targets with the `--target` option into a custom
dirctory under the `target` directory.

I fixed the calcuation by making the path depending on both the selected
target and the `release` flag.

Changes
=======

* ![FIX] cargo complile task output dir based on selected target